requirements.txt
langchain 
openai 
faiss-cpu 
tiktoken 
requests
pdfplumber 
python-dotenv
tavily-python
langchain-openai
langchain-community